requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.3
Django==4.0.1
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
drf-writable-nested==0.6.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
